Issue #7 - Add additional contraints to ffmpeg options
These changes allowed an HD h264 to be successfully transcoded
to a CMDN compaitable mpeg2. This should prevent failed uploads
When the content needs to be transcoded in most cases.

Its possible that the validation that determines what needs to be
transcoded needs to be improved to catch mpegs that don't meet
the  exact specifications of the ACM compaitability. Especially
bitrates, these seem to be the most problematic.